% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamed.R
\name{metamed}
\alias{metamed}
\title{metamed}
\usage{
metamed(
  te,
  te_lb,
  te_ub = NULL,
  de,
  de_lb,
  de_ub = NULL,
  corr = 0.9855538,
  author,
  serv,
  stdserv,
  rr = "exp",
  pmodel = NULL,
  prec = 3,
  weight.prec = 2
)
}
\arguments{
\item{te}{total effect estimates from each individual study}

\item{te_lb}{lower bound of total effect confidence intervals}

\item{te_ub}{upper bound of total effect confidence intervals}

\item{de}{direct effect estimates from each individual study}

\item{de_lb}{lower bound of direct effect confidence intervals}

\item{de_ub}{upper bound of total effect confidence intervals}

\item{corr}{correlation between direct}

\item{author}{label of each individual study}

\item{serv}{serving size of each individual study}

\item{stdserv}{serving size of pooled total effect}

\item{rr}{relative risk function, either "exponential" or identity"}

\item{pmodel}{model used for back-calculated mediation proportion, either
"fixed" or "random" effects model}

\item{prec}{precision of pooled estimates}

\item{weight.prec}{precision of weights on individual studies}
}
\value{
data frames of meta-analysis tables for mediation proportion and
total effect, pooled mediation proportion and total effect results, and
heterogeneity statistics
}
\description{
Estimates a pooled mediation proportion and total effects and
outputs results in data frame
}
\examples{
S <- 30
ST <- 10
SB <- 10
SD <- 10
set.seed(1)
beta.T <- rnorm(ST + SB, mean = 0.3, sd = 0.05)
set.seed(1)
beta.D <- rnorm(SD + SB, mean = 0.2, sd = 0.05)
set.seed(1)
var.beta.T <- rnorm(ST + SB, mean = 0.05, sd = 0.01)
set.seed(1)
var.beta.D <- rnorm(SD + SB, mean = 0.05, sd = 0.01)
df <- data.frame(beta.T = c(beta.T, rep(NA, SD)),
                 beta.D = c(rep(NA, ST), beta.D),
                 var.beta.T = c(var.beta.T, rep(NA, SD)),
                 var.beta.D = c(rep(NA, ST), var.beta.D))
res <- metamed(te = df$beta.T,
               te_lb = df$beta.T - qnorm(0.975) * sqrt(df$var.beta.T),
               de = df$beta.D,
               de_lb = df$beta.D - qnorm(0.975) * sqrt(df$var.beta.D),
               author = paste("Study", seq(1:S)),
               serv = 1,
               stdserv = 1,
               rr = "identity",
               pmodel = "fixed")
}
