% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamed.R
\name{metamed}
\alias{metamed}
\title{metamed}
\usage{
metamed(
  te,
  te_lb,
  te_ub = NULL,
  de,
  de_lb,
  de_ub = NULL,
  corr = 0.9855538,
  author,
  serv,
  stdserv,
  rr = "exp",
  pmodel = NULL,
  prec = 3,
  weight.prec = 2
)
}
\arguments{
\item{te}{total effect estimates from each individual study}

\item{te_lb}{lower bound of total effect confidence intervals}

\item{te_ub}{upper bound of total effect confidence intervals}

\item{de}{direct effect estimates from each individual study}

\item{de_lb}{lower bound of direct effect confidence intervals}

\item{de_ub}{upper bound of total effect confidence intervals}

\item{corr}{correlation between direct}

\item{author}{label of each individual study}

\item{serv}{serving size of each individual study}

\item{stdserv}{serving size of pooled total effect}

\item{rr}{relative risk function, either "exponential" or identity"}

\item{pmodel}{model used for back-calculated mediation proportion, either
"fixed" or "random" effects model}

\item{prec}{precision of pooled estimates}

\item{weight.prec}{precision of weights on individual studies}
}
\value{
data frames of meta-analysis tables for mediation proportion and
total effect, pooled mediation proportion and total effect results, and
heterogeneity statistics
}
\description{
Estimates a pooled mediation proportion and total effects and
outputs results in data frame
}
